requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
django-filter==23.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
